feat : payment, category bad request 처리 기능 추가
update, delete bad request 처리 추가

+++ server/src/domain/payment/payment.service.ts
@@ -1,5 +1,6 @@
 import PaymentEntity from '@/entity/payment.entity';
 import { Repository } from 'typeorm';
+import { BAD_REQUEST } from '@/common/error';
 
 export default class PaymentService {
   private paymentRepository: Repository<PaymentEntity>;
@@ -21,11 +22,17 @@ export default class PaymentService {
     return payment;
   }
 
-  public async updatePayment(paymentId: number, name: string): Promise<void> {
-    await this.paymentRepository.update({ pid: paymentId }, { name });
+  public async updatePayment(paymentId: number, name: string, uid: number): Promise<void> {
+    const { affected } = await this.paymentRepository.update({ pid: paymentId, uid }, { name });
+    if (!affected) {
+      throw new Error(BAD_REQUEST);
+    }
   }
 
-  public async deletePayment(paymentId: number): Promise<void> {
-    await this.paymentRepository.delete({ pid: paymentId });
+  public async deletePayment(paymentId: number, uid: number): Promise<void> {
+    const { affected } = await this.paymentRepository.delete({ pid: paymentId, uid });
+    if (!affected) {
+      throw new Error(BAD_REQUEST);
+    }
   }
 }
